Pass username & password to constructor of the API
In this way, there is no need to get the username & password from the PrefStore
on each API call.

onResume, a new API instance will be created. The rationale behind this is as
follows: On App startup, the sequence will be onCreate -> onResume. If the user
switches to the settings menu, onResume will be called again, and new
credentials are applied.

No need to call checkPreconditions() or getStatus() in onCreate.

Signed-off-by: Ralf Ramsauer <ralf@binary-kitchen.de>