HOTFIX: Move spaceapi URL to R string resources
We had an update fuckup... In 70793d43f ("Allow to specify Space API URL"), we
allow users to specify their own spaceapi URL. This triggers a bug, if the
users directly updates from 2.2 to 2.3: When the app starts, the setting is
missing. It is written to the SharedPreference storage once the settings screen
is opened. Before that, we will get an empty string and the App will crash, if
a widget is active, as it tries to parse an empty URL.

Fix this by falling back to an R resource if we read an empty string.

Fixes: 70793d43f ("Allow to specify Space API URL")
Signed-off-by: Ralf Ramsauer <ralf@binary-kitchen.de>